+++ lib/cache.js
@@ -521,98 +521,20 @@ function cloneGitRemote (p, u, co, origUrl, silent, cb) {
   })
 }
 
+function gitDoneCb (tmp, resolved, cb) {
+  addLocalTarball(tmp, function (er, data) {
+    if (data) data._resolved = resolved
+    cb(er, data)
+  })
+}
+
 function archiveGitRemote (p, u, co, origUrl, cb) {
   var git = npm.config.get("git")
-  var archive = [ "fetch", "-a", "origin" ]
-  var resolve = [ "rev-list", "-n1", co ]
-  var env = gitEnv()
+    , t = npm.tmp
+    , c = npm.cache
 
-  var errState = null
-  var n = 0
-  var resolved = null
-  var tmp
-
-  exec(git, archive, {cwd: p, env: env}, function (er, stdout, stderr) {
-    stdout = (stdout + "\n" + stderr).trim()
-    if (er) {
-      log.error("git fetch -a origin ("+u+")", stdout)
-      return cb(er)
-    }
-    log.verbose("git fetch -a origin ("+u+")", stdout)
-    tmp = path.join(npm.tmp, Date.now()+"-"+Math.random(), "tmp.tgz")
-    verifyOwnership()
-  })
-
-  function verifyOwnership() {
-    if (process.platform === "win32") {
-      log.silly("verifyOwnership", "skipping for windows")
-      resolveHead()
-    } else {
-      getCacheStat(npm.cache, log, function(er, cs) {
-        if (er) {
-          log.error("Could not get cache stat")
-          return cb(er)
-        }
-        chownr(p, cs.uid, cs.gid, function(er) {
-          if (er) {
-            log.error("Failed to change folder ownership under npm cache for %s", p)
-            return cb(er)
-          }
-          resolveHead()
-        })
-      })
-    }
-  }
-
-  function resolveHead () {
-    exec(git, resolve, {cwd: p, env: env}, function (er, stdout, stderr) {
-      stdout = (stdout + "\n" + stderr).trim()
-      if (er) {
-        log.error("Failed resolving git HEAD (" + u + ")", stderr)
-        return cb(er)
-      }
-      log.verbose("git rev-list -n1 " + co, stdout)
-      var parsed = url.parse(origUrl)
-      parsed.hash = stdout
-      resolved = url.format(parsed)
-
-      // https://github.com/npm/npm/issues/3224
-      // node incorrectly sticks a / at the start of the path
-      // We know that the host won't change, so split and detect this
-      var spo = origUrl.split(parsed.host)
-      var spr = resolved.split(parsed.host)
-      if (spo[1].charAt(0) === ':' && spr[1].charAt(0) === '/')
-        spr[1] = spr[1].slice(1)
-      resolved = spr.join(parsed.host)
-
-      log.verbose('resolved git url', resolved)
-      next()
-    })
-  }
-
-  function next () {
-    mkdir(path.dirname(tmp), function (er) {
-      if (er) return cb(er)
-      var gzip = zlib.createGzip({ level: 9 })
-      var git = npm.config.get("git")
-      var args = ["archive", co, "--format=tar", "--prefix=package/"]
-      var out = fs.createWriteStream(tmp)
-      var env = gitEnv()
-      cb = once(cb)
-      var cp = spawn(git, args, { env: env, cwd: p })
-      cp.on("error", cb)
-      cp.stderr.on("data", function(chunk) {
-        log.silly(chunk.toString(), "git archive")
-      })
-
-      cp.stdout.pipe(gzip).pipe(out).on("close", function() {
-        addLocalTarball(tmp, function(er, data) {
-          if (data) data._resolved = resolved
-          cb(er, data)
-        })
-      })
-    })
-  }
+  require("./git-cache.js")
+    .archiveGitRemote(p, u, co, origUrl, c, t, git, log, gitDoneCb, cb)
 }
 
 var gitEnv_
